# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mars Rover API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v2/plateaus": {
            "get": {
                "tags": [
                    "plateau"
                ],
                "operationId": "5b85a65d5454abfd064cd87aed3d640c",
                "responses": {
                    "200": {
                        "description": "Returns all plateaus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "The plateau id",
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "created_at": {
                                                "description": "Created at",
                                                "type": "string",
                                                "example": "2021-08-26T18:41:26.000000Z"
                                            },
                                            "updated_at": {
                                                "description": "Updated at",
                                                "type": "string",
                                                "example": "2021-08-26T18:41:26.000000Z"
                                            },
                                            "width": {
                                                "description": "Plateau width",
                                                "type": "integer",
                                                "example": "50"
                                            },
                                            "height": {
                                                "description": "Plateau height",
                                                "type": "integer",
                                                "example": "50"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "The error description",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "plateau"
                ],
                "operationId": "0a11760645957a349fd44d8c27b42ec3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "x": {
                                        "description": "The plateau width",
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "y": {
                                        "description": "The plateau height",
                                        "type": "integer",
                                        "example": "50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns plateau",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "The plateau id",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "description": "Created at",
                                            "type": "string",
                                            "example": "2021-08-26T18:41:26.000000Z"
                                        },
                                        "updated_at": {
                                            "description": "Updated at",
                                            "type": "string",
                                            "example": "2021-08-26T18:41:26.000000Z"
                                        },
                                        "width": {
                                            "description": "Plateau width",
                                            "type": "integer",
                                            "example": "50"
                                        },
                                        "height": {
                                            "description": "Plateau height",
                                            "type": "integer",
                                            "example": "50"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "The error description",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "The error description",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "The error description",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}